started work with dot icons
