+++ android/src/main/java/io/parity/signer/components/networkicon/blockies/svalinn/BlockiesPainter.kt
@@ -1,4 +1,4 @@
-package io.parity.signer.components.blockies.svalinn
+package io.parity.signer.components.networkicon.blockies.svalinn
 
 
 import androidx.compose.ui.geometry.Offset
